requirements.txt
flask
uv
flask_login
flask_sqlalchemy
flask_migrate
flask_wtf
email_validator